[27] classes view is state only
