+++ BodyPositiveZumba/Classes/ClassesDomain/ClassesViewModel.swift
@@ -8,5 +8,10 @@
 import SwiftUI
 
 class ClassesViewModel: ObservableObject {
-
+    
+    @Published var viewState: ClassesViewState = ClassesViewState()
+    
+    func toggleExpansion() {
+        viewState.isBioExpanded.toggle()
+    }
 }
