+++ BodyPositiveZumba/Massage/MassageUI/MassageView.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 
 struct MassageView: View, ActionableView {
+    
     enum Action {
         case updateUrl
     }
